refactoring id limit

+++ src/main/java/com/sample/urlshortener/service/EncodeService.java
@@ -1,6 +1,6 @@
 package com.sample.urlshortener.service;
 
 public interface EncodeService {
-    String encode(long n);
+    String encode(int n);
     int decode(String s);
 }
